pyproject.toml
# SPDX-FileCopyrightText: Copyright 2022 Canonical Ltd
# SPDX-License-Identifier: GPL-3.0-or-later

[project]
name = "warty"
version = "0.1.0"
description = "Run at your peril!"
authors = [{ name = "Canonical Ltd", email = "no-reply@canonical.com" }]
dependencies = [
    "black",
    "beautifulsoup4",
    "chevron",
    "fabric",
    "github3.py",
    "humanfriendly",  # see also: humanize
    "jira",
    "launchpadlib",
    "mdformat",
    "paramiko==2.8.1",  # FIXME: avoid DSA/RSA confusion (TODO: get URL)
    "pydantic",
    "pymacaroons",
    "python-dotenv",
    "requests",
    "reuse",
    "sh",
]
requires-python = ">=3.10"
license = { text = "GPL-3.0-or-later" }

[project.optional-dependencies]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
skip-magic-trailing-comma = true

[tool.syllabub.files]
body = "static/body.md"
logo = "static/logo.png"

[tool.syllabub.project]
conduct = "ubuntu-2.0"
copyright = "Copyright 2022 Canonical Ltd"
origin = "DCO-1.1-git-signoff"
sbom = "reuse spdx"
title = "Warty scripts 🐗"
url = "https://github.com/lofidevops/warty"
